+++ inc/reception.class.php
@@ -407,10 +407,10 @@ class PluginOrderReception extends CommonDBChild
             echo "<tr class='tab_bg_1 center'>";
             echo "<td></td>";
             echo "<td align='center'>" . $item->getTypeName() . "</td>";
-            echo "<td align='center'>" . Dropdown::getDropdownName(
+            echo "<td align='center'>" . htmlentities(Dropdown::getDropdownName(
                 "glpi_manufacturers",
                 $data_ref["manufacturers_id"],
-            ) . "</td>";
+            ), ENT_QUOTES, "UTF-8") . "</td>";
             if ($table == 'glpi_plugin_order_referencefrees') {
                 echo "<td>" . $data_ref['name'] . "</td>";
             } else {
@@ -576,7 +576,7 @@ class PluginOrderReception extends CommonDBChild
 
                 echo "<td align='center'>" . Dropdown::getDropdownName(getTableForItemType(Entity::class), (int) $data["entities_id"]) . "</td>";
                 echo "<td align='center'>" . Html::convDate($data["delivery_date"]) . "</td>";
-                echo "<td align='center'>" . $data["delivery_number"] . "</td>";
+                echo "<td align='center'>" . htmlentities($data["delivery_number"] ?? "", ENT_QUOTES, "UTF-8") . "</td>";
                 echo "<td align='center'>"
                  . Dropdown::getDropdownName(
                      "glpi_plugin_order_deliverystates",